Added cookies to save chosen story+character
