+++ OstreCWEB/OstreCWEB.Services/Seed/ISeeder.cs
@@ -1,7 +1,13 @@
-﻿namespace OstreCWEB.Data.DataBase
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OstreCWEB.Services.Seed
 {
     public interface ISeeder
     {
-        public Task SeedDataBase(); 
+        public Task SeedCharactersDb();
     }
 }
