+++ OstreCWEB/OstreCWEB/Controllers/GameController.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Options;
 using OstreCWEB.Data.Interfaces;
 using OstreCWEB.Services.Characters;
 using OstreCWEB.Services.Identity;
@@ -9,6 +10,7 @@ using OstreCWEB.ViewModel.Characters;
 using OstreCWEB.ViewModel.Game;
 using OstreCWEB.ViewModel.Identity;
 using OstreCWEB.ViewModel.StoryBuilder;
+using System.Net;
 
 namespace OstreCWEB.Controllers
 {
@@ -18,17 +20,26 @@ namespace OstreCWEB.Controllers
         private readonly IMapper _mapper;
         private readonly IStoryService _storyService;
         private readonly IPlayableCharacterService _playableCharacterService;
-        public GameController(IUserService userService, IMapper mapper, IStoryService storyService, IPlayableCharacterService playableCharacterService)
+        private readonly IHttpContextAccessor _httpContextAccessor;
+
+        public GameController(IHttpContextAccessor httpContextAccessor,IUserService userService, IMapper mapper, IStoryService storyService, IPlayableCharacterService playableCharacterService)
         {
             _userService = userService;
             _mapper = mapper;
             _storyService = storyService;
             _playableCharacterService = playableCharacterService;
+            _httpContextAccessor = httpContextAccessor;
         }
 
         // GET: GameController
-        public async Task<ActionResult> Index()
+        [HttpPost]
+        public async Task<ActionResult> StartGame(StartGameView model)
         {
+            //Redirets to story reader with character + story chosen by user.
+            return RedirectToAction("Index","StoryReader");
+        }
+        public async Task<ActionResult> Index()
+        { 
             var model = new StartGameView(); 
 
             model.User =  _mapper.Map<UserView>(await _userService.GetUserById(_userService.GetUserId(User))); 
@@ -43,11 +54,37 @@ namespace OstreCWEB.Controllers
             {
                 var mappedCharacter = _mapper.Map<PlayableCharacterRow>(character);
                 model.OtherUsersCharacters.Add(mappedCharacter);
-            }  
+            } 
             return View(model);
         }
+        [HttpGet]
+        public ActionResult SetActiveStory(int id)
+        { 
+            CookieOptions options = new CookieOptions();
+            options.Expires = DateTime.Now.AddSeconds(10);
+            options.Path = "/"; 
+            //Bypasses consent policy checks.
+            options.IsEssential=true;   
+            _httpContextAccessor.HttpContext.Response.Cookies.Append("ActiveStory", $"{id}", options);
 
+            return RedirectToAction(nameof(Index));
+           
+        }
+        [HttpGet]
+        public ActionResult SetActiveCharacter(int id)
+        {
+            CookieOptions options = new CookieOptions();
+            options.Expires = DateTime.Now.AddSeconds(10);
+            options.Path = "/";
+            //Bypasses consent policy checks.
+            options.IsEssential = true;
+            _httpContextAccessor.HttpContext.Response.Cookies.Append("ActiveCharacter", $"{id}", options);
+
+            return RedirectToAction(nameof(Index));
+
+        }
         // GET: GameController/Details/5
+
         public ActionResult Details(int id)
         {
             return View();
